implement query API
small refactor by moving SynqStub to api_test.go
