+++ sample/query.json
@@ -1,0 +1,131 @@
+[
+  {
+    "input": {
+      "url": "https://multicdn.synq.fm/projects/0a/bf/0abfe1b849154082993f2fce77a16fd9/uploads/videos/45/d4/45d4063d00454c9fb21e5186a09c3115.mp4",
+      "width": 720,
+      "height": 1280,
+      "duration": 17.48,
+      "file_size": 16706384,
+      "framerate": 29.97,
+      "uploaded_at": "2017-02-15T03:05:17.978Z"
+    },
+    "state": "uploaded",
+    "player": {
+      "views": 2,
+      "embed_url": "https://player.synq.fm/embed/45d4063d00454c9fb21e5186a09c311",
+      "thumbnail_url": "https://multicdn.synq.fm/projects/0a/bf/0abfe1b849154082993f2fce77a16fd9/derivatives/thumbnails/45/d4/45d4063d00454c9fb21e5186a09c3115/0000360.jpg"
+    },
+    "outputs": {
+      "hls": {
+        "url": "https://multicdn.synq.fm/projects/0a/bf/0abfe1b849154082993f2fce77a16fd9/derivatives/videos/45/d4/45d4063d00454c9fb21e5186a09c3115/hls/45d4063d00454c9fb21e5186a09c3115_hls.m3u8",
+        "state": "complete"
+      },
+      "mp4_360": {
+        "url": "https://multicdn.synq.fm/projects/0a/bf/0abfe1b849154082993f2fce77a16fd9/derivatives/videos/45/d4/45d4063d00454c9fb21e5186a09c3115/mp4_360/45d4063d00454c9fb21e5186a09c3115_mp4_360.mp4",
+        "state": "complete"
+      },
+      "mp4_720": {
+        "url": "https://multicdn.synq.fm/projects/0a/bf/0abfe1b849154082993f2fce77a16fd9/derivatives/videos/45/d4/45d4063d00454c9fb21e5186a09c3115/mp4_720/45d4063d00454c9fb21e5186a09c3115_mp4_720.mp4",
+        "state": "complete"
+      },
+      "mp4_1080": {
+        "url": "https://multicdn.synq.fm/projects/0a/bf/0abfe1b849154082993f2fce77a16fd9/derivatives/videos/45/d4/45d4063d00454c9fb21e5186a09c3115/mp4_1080/45d4063d00454c9fb21e5186a09c3115_mp4_1080.mp4",
+        "state": "complete"
+      },
+      "webm_720": {
+        "url": "https://multicdn.synq.fm/projects/0a/bf/0abfe1b849154082993f2fce77a16fd9/derivatives/videos/45/d4/45d4063d00454c9fb21e5186a09c3115/webm_720/45d4063d00454c9fb21e5186a09c3115_webm_720.webm",
+        "state": "complete"
+      }
+    },
+    "userdata": {},
+    "video_id": "45d4063d00454c9fb21e5186a09c3115",
+    "created_at": "2017-02-15T03:01:16.767Z",
+    "updated_at": "2017-02-15T03:06:31.794Z"
+  },
+  {
+    "input": {
+      "url": "https://multicdn.synq.fm/projects/0a/bf/0abfe1b849154082993f2fce77a16fd9/uploads/videos/de/98/de98a4c92152411fbac3b7027c8f2df7.mp4",
+      "width": 1152,
+      "height": 720,
+      "duration": 353.4,
+      "file_size": 127589697,
+      "framerate": 32.37,
+      "uploaded_at": "2017-03-08T20:45:58.784Z"
+    },
+    "state": "uploaded",
+    "player": {
+      "views": 2,
+      "embed_url": "https://player.synq.fm/embed/de98a4c92152411fbac3b7027c8f2df7",
+      "thumbnail_url": "https://multicdn.synq.fm/projects/0a/bf/0abfe1b849154082993f2fce77a16fd9/derivatives/thumbnails/de/98/de98a4c92152411fbac3b7027c8f2df7/0009979.jpg"
+    },
+    "outputs": {
+      "hls": {
+        "url": "https://multicdn.synq.fm/projects/0a/bf/0abfe1b849154082993f2fce77a16fd9/derivatives/videos/de/98/de98a4c92152411fbac3b7027c8f2df7/hls/de98a4c92152411fbac3b7027c8f2df7_hls.m3u8",
+        "state": "complete"
+      },
+      "mp4_360": {
+        "url": "https://multicdn.synq.fm/projects/0a/bf/0abfe1b849154082993f2fce77a16fd9/derivatives/videos/de/98/de98a4c92152411fbac3b7027c8f2df7/mp4_360/de98a4c92152411fbac3b7027c8f2df7_mp4_360.mp4",
+        "state": "complete"
+      },
+      "mp4_720": {
+        "url": "https://multicdn.synq.fm/projects/0a/bf/0abfe1b849154082993f2fce77a16fd9/derivatives/videos/de/98/de98a4c92152411fbac3b7027c8f2df7/mp4_720/de98a4c92152411fbac3b7027c8f2df7_mp4_720.mp4",
+        "state": "complete"
+      },
+      "mp4_1080": {
+        "url": "https://multicdn.synq.fm/projects/0a/bf/0abfe1b849154082993f2fce77a16fd9/derivatives/videos/de/98/de98a4c92152411fbac3b7027c8f2df7/mp4_1080/de98a4c92152411fbac3b7027c8f2df7_mp4_1080.mp4",
+        "state": "complete"
+      },
+      "webm_720": {
+        "url": "https://multicdn.synq.fm/projects/0a/bf/0abfe1b849154082993f2fce77a16fd9/derivatives/videos/de/98/de98a4c92152411fbac3b7027c8f2df7/webm_720/de98a4c92152411fbac3b7027c8f2df7_webm_720.webm",
+        "state": "complete"
+      }
+    },
+    "userdata": {},
+    "video_id": "de98a4c92152411fbac3b7027c8f2df7",
+    "created_at": "2017-03-08T20:11:08.199Z",
+    "updated_at": "2017-03-08T20:56:09.017Z"
+  },
+  {
+    "input": {
+      "url": "https://multicdn.synq.fm/projects/0a/bf/0abfe1b849154082993f2fce77a16fd9/uploads/videos/bb/98/bb98e6cea8224ea29e6bf00e36632bdf.mp4",
+      "width": 1280,
+      "height": 720,
+      "duration": 415.4,
+      "file_size": 127438343,
+      "framerate": 27.38,
+      "uploaded_at": "2017-05-02T16:35:18.977Z"
+    },
+    "state": "uploaded",
+    "player": {
+      "views": 2,
+      "embed_url": "https://player.synq.fm/embed/bb98e6cea8224ea29e6bf00e36632bdf",
+      "thumbnail_url": "https://multicdn.synq.fm/projects/0a/bf/0abfe1b849154082993f2fce77a16fd9/derivatives/thumbnails/bb/98/bb98e6cea8224ea29e6bf00e36632bdf/0008183.jpg"
+    },
+    "outputs": {
+      "hls": {
+        "url": "https://multicdn.synq.fm/projects/0a/bf/0abfe1b849154082993f2fce77a16fd9/derivatives/videos/bb/98/bb98e6cea8224ea29e6bf00e36632bdf/hls/bb98e6cea8224ea29e6bf00e36632bdf_hls.m3u8",
+        "state": "complete"
+      },
+      "mp4_360": {
+        "url": "https://multicdn.synq.fm/projects/0a/bf/0abfe1b849154082993f2fce77a16fd9/derivatives/videos/bb/98/bb98e6cea8224ea29e6bf00e36632bdf/mp4_360/bb98e6cea8224ea29e6bf00e36632bdf_mp4_360.mp4",
+        "state": "complete"
+      },
+      "mp4_720": {
+        "url": "https://multicdn.synq.fm/projects/0a/bf/0abfe1b849154082993f2fce77a16fd9/derivatives/videos/bb/98/bb98e6cea8224ea29e6bf00e36632bdf/mp4_720/bb98e6cea8224ea29e6bf00e36632bdf_mp4_720.mp4",
+        "state": "complete"
+      },
+      "mp4_1080": {
+        "url": "https://multicdn.synq.fm/projects/0a/bf/0abfe1b849154082993f2fce77a16fd9/derivatives/videos/bb/98/bb98e6cea8224ea29e6bf00e36632bdf/mp4_1080/bb98e6cea8224ea29e6bf00e36632bdf_mp4_1080.mp4",
+        "state": "complete"
+      },
+      "webm_720": {
+        "url": "https://multicdn.synq.fm/projects/0a/bf/0abfe1b849154082993f2fce77a16fd9/derivatives/videos/bb/98/bb98e6cea8224ea29e6bf00e36632bdf/webm_720/bb98e6cea8224ea29e6bf00e36632bdf_webm_720.webm",
+        "state": "complete"
+      }
+    },
+    "userdata": {},
+    "video_id": "bb98e6cea8224ea29e6bf00e36632bdf",
+    "created_at": "2017-05-02T16:24:26.349Z",
+    "updated_at": "2017-05-02T19:07:10.065Z"
+  }
+]
